% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.cluster_stability}
\alias{.cluster_stability}
\title{Resampling-based stability assessment using ARI}
\usage{
.cluster_stability(
  S,
  times,
  B = 30,
  frac = 0.7,
  mode = c("subsample", "bootstrap"),
  jitter_sd = 0.001,
  weight_perturb = 0.001,
  seed = NULL,
  return_distribution = TRUE,
  ...
)
}
\description{
Resampling-based stability assessment using ARI
}
\keyword{internal}
