% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.enforce_monotone}
\alias{.enforce_monotone}
\title{Enforce non-increasing survival curves over time}
\usage{
.enforce_monotone(S)
}
\description{
Enforce non-increasing survival curves over time
}
\keyword{internal}
