% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.smooth_median}
\alias{.smooth_median}
\title{Median smooth survival curves along the time axis}
\usage{
.smooth_median(S, width)
}
\description{
Median smooth survival curves along the time axis
}
\keyword{internal}
