% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.weight_features}
\alias{.weight_features}
\title{Apply time-point weights to curve features}
\usage{
.weight_features(S, w, distance)
}
\description{
Apply time-point weights to curve features
}
\keyword{internal}
