% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.unsurv}
\alias{plot.unsurv}
\title{Plot medoid survival curves from an unsurv fit}
\usage{
\method{plot}{unsurv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"unsurv"}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{matplot}}
(e.g., \code{lwd}).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Produces a base R plot of the cluster medoid survival curves stored in the
fitted object.
}
