% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{plot_surv_medoids}
\alias{plot_surv_medoids}
\title{Plot medoid curves with ggplot2 (internal helper)}
\usage{
plot_surv_medoids(fit)
}
\description{
Plot medoid curves with ggplot2 (internal helper)
}
\keyword{internal}
