% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{plot_surv_samples}
\alias{plot_surv_samples}
\title{Plot individual curves with ggplot2 (internal helper)}
\usage{
plot_surv_samples(S, times, clusters = NULL, alpha = 0.2)
}
\description{
Plot individual curves with ggplot2 (internal helper)
}
\keyword{internal}
