% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print.summary.unsurv}
\alias{print.summary.unsurv}
\title{Print a summary of an unsurv model}
\usage{
\method{print}{summary.unsurv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.unsurv"}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of an unsurv model
}
