% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print.unsurv}
\alias{print.unsurv}
\title{Print an unsurv model}
\usage{
\method{print}{unsurv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"unsurv"}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an unsurv model
}
