% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.unsurv_compare}
\alias{print.unsurv_compare}
\title{Print a partition comparison}
\usage{
\method{print}{unsurv_compare}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"unsurv_compare"}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a partition comparison
}
