% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{autoplot.unsurv}
\alias{autoplot.unsurv}
\title{ggplot2 autoplot for unsurv objects}
\usage{
\method{autoplot}{unsurv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"unsurv"}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
ggplot2 autoplot for unsurv objects
}
