% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{plot_stability}
\alias{plot_stability}
\title{Plot stability distribution (internal helper)}
\usage{
plot_stability(stab)
}
\description{
Plot stability distribution (internal helper)
}
\keyword{internal}
