% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary.unsurv}
\alias{summary.unsurv}
\title{Summarize an unsurv model}
\usage{
\method{summary}{unsurv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"unsurv"}.}

\item{...}{Unused.}
}
\value{
An object of class \code{"summary.unsurv"} with elements:
\itemize{
\item \code{K}: number of clusters
\item \code{silhouette_mean}: mean silhouette width
\item \code{size}: cluster sizes
}
}
\description{
Summarize an unsurv model
}
